requirements.txt
playwright>=1.41.0,<2.0.0
pytest>=8.2,<9.0
pytest-asyncio>=0.23.0,<1.0.0
allure-pytest>=2.15.0,<3.0.0
pytest-playwright-asyncio>=0.7.0,<1.0.0
pytest-xdist>=3.0.0,<4.0.0
